tractorbeam/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
loguru==0.7.2
mako==1.3.2
markupsafe==2.1.5
packaging==23.2
pluggy==1.4.0
pydantic==2.6.3
pydantic-core==2.16.3
pydantic-settings==2.2.1
pyjwt==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.5.post1
python-dotenv==1.0.1
requests==2.31.0
ruff==0.3.2
sniffio==1.3.1
sqlalchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
typing-extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
